main.py
import time, cv2

from ppocrv5_ov.predict_system import TextSystem
from ppocrv5_ov.utils import infer_args as init_args
from ppocrv5_ov.utils import str2bool, draw_ocr
import argparse

class PPOCRv4_OV(TextSystem):
    def __init__(self, args):
        self.args = args
        # Initialize all the models
        super().__init__(args)

    def ocr(self, img, det=True, rec=True, cls=True):
        if cls == True and self.args.use_angle_cls == False:
            print(
                "Since the angle classifier is not initialized, the angle classifier will not be uesd during the forward process"
            )

        if det and rec:
            ocr_res = []
            dt_boxes, rec_res = self.__call__(img, cls)
            tmp_res = [[box.tolist(), res] for box, res in zip(dt_boxes, rec_res)]
            ocr_res.append(tmp_res)
            return ocr_res
        elif det and not rec:
            ocr_res = []
            dt_boxes = self.text_detector(img)
            tmp_res = [box.tolist() for box in dt_boxes]
            ocr_res.append(tmp_res)
            return ocr_res
        else:
            ocr_res = []
            cls_res = []

            if not isinstance(img, list):
                img = [img]
            if self.args.use_angle_cls and cls:
                img, cls_res_tmp = self.text_classifier(img)
                if not rec:
                    cls_res.append(cls_res_tmp)
            rec_res = self.text_recognizer(img)
            ocr_res.append(rec_res)

            if not rec:
                return cls_res
            return ocr_res


def sav2Img(org_img, result, name="draw_ocr.jpg"):
    # 显示结果
    from PIL import Image

    result = result[0]
    image = org_img[:, :, ::-1]
    boxes = [line[0] for line in result]
    txts = [line[1][0] for line in result]
    scores = [line[1][1] for line in result]
    im_show = draw_ocr(image, boxes, txts, scores)
    im_show = Image.fromarray(im_show)
    im_show.save(name)


def main(args):
    # Change the default values of the arguments in the utils.infer_args function
    model = PPOCRv4_OV(args)

    img = cv2.imread(args.image_dir)

    s = time.time()
    result = model.ocr(img)
    e = time.time()
    print("total time: {:.3f}".format(e - s))
    print("result:", result)
    for box in result[0]:
        print(box)

    sav2Img(img, result)


if __name__ == "__main__":
    
    parser = init_args()
    args = parser.parse_args()

    main(args)
