requirements.txt
openvino
opencv-python
Pillow
shapely
pyclipper
scikit-image
imgaug
lmdb
tqdm